reduce max fix size
